Adds a waitFor helper for asset criticality records
We were observing a synchronization issue while developing the API
integration tests: criticality records were not immediately available
for search. This is not unexpected or unusual, but the existing `get()`
function seems to have bypassed this since it queries directly by `_id`
and doesn't need to wait for an index refresh.

The diff here is messy because this also deletes some old, unused
helpers from this branch.