Fix flaky test around risk task execution
The issue here was that be our assertion that the number of generated
risk scores be EXACTLY 20; in a case where the test runner itself is
slow (but kibana/ES is not), the task can/may execute more than one
time. In the accompanying issue we can see that that's indeed what
happened: the task executed four times, generating 80 alerts, and our
assertion failed.

Closes #171132.