[Metric threshold] Fix incorrect rule params when changing nonCount agg to count in custom equation (#171958)
Closes #171088

## Summary

This PR fixes the incorrect params when changing aggregation from
nonCount to count:


https://github.com/elastic/kibana/assets/12370520/56ae5612-e254-4815-98fa-e773cbe4ba38

I didn't find an easy way to add a test for this case, as this is a
minor UI issue. It is worth mentioning that the metric threshold rule
will be deprecated eventually. Let me know if you have any suggestions
for it.