[ML] [AIOps] Using standard analyser in pattern analysis (#172188)
Fixes https://github.com/elastic/kibana/issues/169523

The `categorize_text` agg uses the `ml_standard` tokenizer by default
which produces slightly different tokens compared to the `standard`
tokenizer, which is the default used for search.
This means the category key (which is comprised of these tokens) will
occasionally not match any documents when it is used as a filter in
Discover to find docs in a category.

This PR ensures the `standard` tokenizer is always used in the pattern
analysis query.

A future enhancement would be to check which analyzer is specified in
the mappings for the source field and to use that instead of
unconditionally using `standard`. However for an initial fix, using the
`standard` analyzer will be more likely to match the results from the
majority of searches.