[Security Solution][Investigations] - Fix ESQL query size (#171789)
## Summary

This PR improves a potential performance issue for timeline currently.
When users load a security solution page with timeline, the default ESQL
tab query is run in the background as the query is a part of the url
state. While this can be convenient for the speed of retrieval when
opening the ESQL tab, it can lead to some performance issues on initial
load depending on the number of fields/size of the data in the user's
default security indices. To improve the performance of this initial
query, we'll limit the query to only initially retrieve the default
Timeline columns via the ESQL `keep` clause.

`keep @timestamp, message, event.category, event.action, host.names,
source.ip, destination.ip, user.name` will be appended to the end of the
current default query

<img width="1680" alt="image"
src="https://github.com/elastic/kibana/assets/17211684/5fc3ae69-8ae6-4143-8f75-7cf69032e63a">

**Regarding the unskipped test:**

The tests were flaky due to the fact that interacting with the monaco
editor with the cypress `type` commands didn't allow for deleting of the
existing text, but this was able to be remedied by first expanding the
editor which allows the text to be more easily interacted with via the
cypress commands.

[Buildkite (x100) test
run](https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/4174#_)
- The one failure is due to a flaky test in an unrelated correlation
tab.

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit 6349d1c5131218dc8059616d9a42551d0c0d667c)

# Conflicts:
#	x-pack/plugins/security_solution/public/common/components/discover_in_timeline/use_discover_in_timeline_actions.tsx