[Obs AI Assistant] Abort controller when component unmounts (#172870)
When the component unmounts (for instance when moving to a different
chat or navigating to another page), the request to /complete should be
aborted. This PR ensures that that is the case and adds a test for it.

(cherry picked from commit c0ab01aebc90c11bc868b0115d5bc62cecc8abe9)