Dynamically set pipeline selection list height (#172816)
## Summary

Update the pipeline selection list to set a maximum height that
decreases when there are too few options to fill the space. The maximum
height is set to 4.5 rows, with the extra half row used to indicate to
the user that there are more options available than what is visible
(i.e. they need to scroll). If 4 or fewer options are available, the height is
set to `<option_count>` rows.

(cherry picked from commit 7f61e2a41f1281c27bf82be9379b81a2b7ad80f4)