[Advanced Settings] Fix defaultIndex setting's default value (#170865)
Fixes https://github.com/elastic/kibana/issues/168732

## Summary

This PR changes the default value of the `defaultIndex` setting to an
empty string. Previously, the default value was set to `null` and this
was causing a wrong behaviour in the Advanced Settings UI as it expects
a string value for the settings of type `string`.

**Before - the "Reset to default" link did not clear the user value:**

https://github.com/elastic/kibana/assets/59341489/2c47306b-feac-4e08-bbae-3db7b0a3e5fd

**With these changes:**

https://github.com/elastic/kibana/assets/59341489/dc46caa8-9860-4b66-87ca-9c184c5a8391
(cherry picked from commit 946b38ed577ddfccf1fb022d7fa1ff6e9cdd1606)