Extract risk level constants to files under EA
These started as a way to assist Threat Hunting with their UI. Now that
we need to do the same level calculation during risk scoring, we also
have need for these concepts (namely: a list of levels, and a map of
their ranges).

To keep the changes in Threat Hunting code to a minimum, I opted to just
re-export these in their old location from their new one.