progress, but not complete
To use this, you'll need to use an index threshold rule with
a slack-api connector.  In the slack message for the rule
action, add the string `{{{context.chartData}}}`, which should
add an image to the generated message.

But it's not actually working.  At the point where we try to
generate the PNG data for the chart, we're getting this from
vega / node-canvas:

    Error: CanvasRenderer is missing a valid canvas or context

In theory, I think everything else should work.  I may see if
I can use an alternate library to generate an image, or see
if I can use puppeteer or something instead.