Category scores have the same precision as total risk scores
Now that we are normalizing category scores, the sum of all category
scores should sum to the total normalized risk score. However, it
wasn't, and it was due to the category scores being treated as integers
in our reduce script. By initializing those variables to 0.0, the
interpreter gives them the proper type/treatment.