Update FTR tests with new risk engine updates
There are two main changes reflected here:

1. The addition of asset criticality fields and scoring logic
2. The normalization of category scores

The tests around weighting show the interaction between weighting,
criticality, and normalization: as it stands if a weight has been
applied, the category scores are not themselves weighted (and so will
not sum to the total score). I'll discuss this with the rest of the
team.

There's also some strangeness with the task tests, where the transform
will fail and cause the test server to crash. I'm adding it to the TODO
list.