chore: re-resolve transitive dependency that breaks on CI (#173254)
## Summary
Several cases of the cache population build is failing due to 401 on
`https://packages.atlassian.com/api/npm/npm-remote`.

It seems it's the only one artifact with that resolved path. After a
forced re-resolution, this path goes away, so let's ditch it!

It also seems to be present in `8.12` - so needs a backport

See:
-
https://buildkite.com/elastic/kibana-macos-bazel-cache/builds/29364#018c6293-9782-4b89-aa59-6344ee6fe3a8
-
https://buildkite.com/elastic/kibana-macos-bazel-cache/builds/29365#018c62a9-42c9-49b6-86ae-48f128131bbb
-
https://buildkite.com/elastic/kibana-macos-bazel-cache/builds/29366#018c62b7-776a-4239-94cf-065089a452ba

(cherry picked from commit 5793034b109789fc6acba562f46e6a450daad973)