Update configuring-reporting.asciidoc
A customer opened a case about the steps listed in https://www.elastic.co/guide/en/kibana/8.11/secure-reporting.html#grant-user-access

Per the customer, and I tend to agree, the existing wording is accurate but hard to follow, especially for a non-native english speaker.

In particular, this bit:

"""
Enable application privileges in Reporting. To enable, turn off the default user access control features in kibana.yml:
"""

It's technically accurate, but to dissect the "To enable, turn off..." part you have to do some further reading. I was hoping that maybe there's a way to just rephrase this part or slightly restructure it to make it clear that:

In order to enable application privilege reporting...you must disable the other setting first, otherwise you can't see it. That's my pre-coffee, non-technical-write version.

Please feel free to modify this further if my change is lacking.