[Security Solution] Fixes ML Security Job fetch failure when `securitySolution:defaultIndex` contains special characters (#173426)
## Summary

It was reported that some `securitySolution:defaultIndex`'s could result
in the Security Job request failing. This was identified to be an issue
with the query params not being properly encoded. This PR updates the
`ml_popover` api requests to call `encodeURIComponent()` for all
variables used within the request URL's.

## Test instructions

Set your `securitySolution:defaultIndex` to be something that includes
special characters that wouldn't be encoded in the URL, e.g.:

```
.internal.alerts-security.alerts-siem-, <winlogbeat-{now/d-1d{yyyy.MM.dd}}>, <winlogbeat-{now/d{yyyy.MM.dd}}*>
```

Then navigate to a page that has the `ML job settings` UI (which fetches
the Security ML Jobs), and verify that no errors are displayed. You can
open the network panel and ensure that the URL params are correctly
encoded, and that the request was successful:

<img width="2284" alt="Detection_rules__SIEM__-_Kibana"
src="https://github.com/elastic/kibana/assets/2946766/1e7d039d-3248-409b-acf4-e65e67808b4c">

Additionally, enable a job and ensure it enables successfully.

(cherry picked from commit 51c506a28c519ad74032d660854eb0c3bcbb4987)