Writing of new risk score fields is controlled by feature flag
This is needed because we don't currently have a mechanism to update the
mappings for the risk score indices, and the downstream transform will
crash ES if those fields exist on docs in that index (mappings or no).

Once the above is resolved, we can basically just revert this commit.

This reuses the existing feature flag,
`entityAnalyticsAssetCriticalityEnabled`, to control whether our risk
engine will calculate and write these new fields.

It does this by placing knowledge of that flag inside the
AssetCriticalityService, and modifying behavior based on that within the
`calculateRiskScores` function.

In terms of types, the new fields have all been marked as optional. I
have not modified the API schema because a) they reflect the full
possible field list, and b) I don't believe anything is
currently consuming them.