[Obs AI Assistant] ES|QL improvements (#173303)
Use a combination of system prompt + recall for ES|QL to allow for:

- faster responses
- wider support of functions (e.g. any documented function is now
supported, in theory)

Performance for the current ES|QL scenarios is similar.

Some things I've noticed:
- ~for whatever reason it is much harder to get the LLM to not make
certain mistakes. I've tried mentioning some mistakes close to the end
of the conversation, but will explore other options.~ I am again opting
for repeating common mistakes at the start of the message, but this time
I'm allowing the LLM to pick them.
- if we can prevent it from making "known" mistakes, we can figure out a
way to have the LLM store known mistakes. This is also useful for other
functions
- another thing I want to figure out is if more examples improves for
commands and custom documentation improves performance. but later.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>