Skip an expectedly failing test
This test will fail because it relies on elasticsearch not crashing
while multiple task runs execute. However, in our current situation, any
test that exercises the new criticality code paths will cause kibana to
crash once a) the data is written, and then b) the transform runs on
that new data.

Since these are just integration tests that are run in
isolation, I'm not worried about the "fail after assertion" behavior
elsewhere, but this test needs to be skipped because of it, since it
will never pass until we fix the transform issue (which, again:
https://github.com/elastic/security-team/issues/8012).