[ci/pr] Remove renovate from allowed_list (#173285)
Renovate pull requests often sit inactive before a developer is able to
review. While inactive renovate is adding new versions, or the pull
request becomes outdated and CI cycles are not used.

This moves renovate CI on pull requests to a manual trigger, either via
`buildkite test this` or `/ci`.