[Obs AI Assistant] Use Mocha & Synthtrace (#173993)
Two changes here:

1. Use Mocha instead of custom test runner. This allows us to use
Mocha's much more extensive API to run the tests: skipping tests, before
and after hooks, grep etc.
2. Make Synthtrace available for running tests.

One note:
- I've added a rule to automatically inject a type reference for Mocha,
via `@kbn/ambient-ftr-types`. The reason is Mocha is not typed by
default (ie, the types are not available in the global space), because
it conflicts with Jest. The rule enforces that `@kbn/ambient-ftr-types`
is imported so things like `before()` can be used.