Handle content stream errors in report pre-deletion (#173792)
Re-addresses https://github.com/elastic/kibana/issues/171363

The bug was still evident, especially when using network throttling to
add slight lag to the request turnaround times.

This PR adds more handling of errors that could be thrown slightly prior
to deleting the report document, when we try to clear all chunks of the
report using the content stream.

<details>
<summary>Before</summary>

https://github.com/elastic/kibana/assets/908371/c27fe314-0f93-42b4-8076-99a1e30b8d2f

</details>

<details>
<summary>After</summary>

https://github.com/elastic/kibana/assets/908371/4c1f5edd-73f1-4ca4-a40a-f900ca5f9c78

</details>

### Checklist
- [x] Unit tests